Move NewSearchPage and NewSearchPageTest to package 'io.github.openequella'.
